Dump the current work for DispmanX-based WaylandCompositor
Committing the current set of changes that brings preliminary
support for accelerated compositing in WebKitGTK+ under Wayland
for the Raspberry Pi.

Currently the implemented Wayland compositor constructs an
absolutely positioned window and renders the content into that.
The Wayland protocol design, to my knowledge, doesn't enable us
positioning that window relatively to the parent GtkWidget, but
this might not be a serious problem since the window is planned
to be full-screened.

Again, this is just a code dump that enables the interested people
to test out the current progress, so there are plenty of bugs and
optimizations that still need ironing.